+++ src/BackgroundJobs/JobsManager.cs
@@ -60,7 +60,6 @@ public static class JobsManager
                  x => x.Run(), @"0 */2 * * *"); 
          
          //DEBUG
-         
          //RecurringJob.AddOrUpdate<BackupWork>("DebugService", x => x.Run(), @"*/1 * * * *"); 
             
     }