+++ src/BackgroundJobs/Program.cs
@@ -2,6 +2,8 @@
 
 using System.Runtime.InteropServices;
 using BackgroundJobs;
+using BackgroundJobs.Jobs;
+using BackgroundJobs.Jobs.Calculation;
 using Hangfire;
 using Hangfire.Logging;
 using Hangfire.Logging.LogProviders;
@@ -11,6 +13,8 @@ using Microsoft.Extensions.DependencyInjection;
 using Serilog;
 using Serilog.Core;
 using Serilog.Events;
+using ServerServices.Interfaces;
+using ServerServices.Services;
 using Spectre.Console;
 using ConfigurationManager = BackgroundJobs.ConfigurationManager;