+++ src/ServerServices/Interfaces/IRiskCalculationService.cs
@@ -1,0 +1,12 @@
+using ServerServices.Events;
+
+namespace ServerServices.Interfaces;
+
+public interface IRiskCalculationService
+{
+    
+    public event EventHandler<RiskCalculationEventArgs> RiskScoreCalculated;
+    
+    public Task CalculateRiskScoreAsync();
+    public Task CalculateContributingImpactAsync();
+}