+++ src/ServerServices/Services/RiskCalculationService.cs
@@ -74,10 +74,10 @@ public class RiskCalculationService(
             // Get Risks with vulnerabilities
             var risks = context.Risks
                 .Include(risk => risk.Vulnerabilities.Where(v => 
-                    v.Status != (int)IntStatus.Closed ||
-                    v.Status != (int)IntStatus.Solved ||
-                    v.Status != (int)IntStatus.Rejected ||
-                    v.Status != (int)IntStatus.Retired ||
+                    v.Status != (int)IntStatus.Closed &&
+                    v.Status != (int)IntStatus.Solved &&
+                    v.Status != (int)IntStatus.Rejected &&
+                    v.Status != (int)IntStatus.Retired &&
                     v.Status != (int)IntStatus.Fixed
                 ))
                 .Where(r => r.Vulnerabilities.Count > 0)