+++ src/API/Controllers/FaceIDController.cs
@@ -1,5 +1,6 @@
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using Model.Exceptions;
 using Model.Services;
 using ServerServices.Interfaces;
 using ILogger = Serilog.ILogger;
@@ -12,14 +13,14 @@ namespace API.Controllers;
 [Route("[controller]")]
 public class FaceIDController: ApiBaseController
 {   
-    private readonly IFaceIDService _faceIDService;
+    private  IFaceIDService FaceIDService { get; }
     
     public FaceIDController(ILogger logger,
         IHttpContextAccessor httpContextAccessor,
         IFaceIDService faceIDService,
         IUsersService usersService) : base(logger, httpContextAccessor, usersService)
     {
-        _faceIDService = faceIDService;
+        FaceIDService = faceIDService;
     }
     
     [HttpGet]
@@ -27,7 +28,7 @@ public class FaceIDController: ApiBaseController
     public async Task<ActionResult<ServiceInformation>> GetInfo()
     {
         
-        return await _faceIDService.GetInfoAsync();
+        return await FaceIDService.GetInfoAsync();
 
     }
     
@@ -40,8 +41,22 @@ public class FaceIDController: ApiBaseController
     [Route("enabled/{userId}")]
     public async Task<ActionResult<bool>> CheckUserEnabled(int userId)
     {
+        try
+        {
+            var result = await FaceIDService.IsUserEnabledAsync(userId);
         
-        return false;
+            return result;
+            
+        }catch (UserNotFoundException e)
+        {
+            return NotFound(e.Message);
+        }
+        catch (Exception e)
+        {
+            Logger.Error(e, "Error checking user enabled status");
+            return StatusCode(500, "Internal server error");
+        }
+
 
     }
     
@@ -54,8 +69,22 @@ public class FaceIDController: ApiBaseController
     [Route("enable/{userId}")]
     public async Task<ActionResult> EnableUser(int userId)
     {
+        try
+        {
+            await FaceIDService.SetUserEnabledStatusAsync(userId, true);
         
-        return Ok();
+            return Ok();
+            
+        }catch (UserNotFoundException e)
+        {
+            return NotFound(e.Message);
+        }
+        catch (Exception e)
+        {
+            Logger.Error(e, "Error enabling user faceid status");
+            return StatusCode(500, "Internal server error");
+        }
+
 
     }
     
@@ -69,7 +98,21 @@ public class FaceIDController: ApiBaseController
     public async Task<ActionResult> DisableUser(int userId)
     {
         
-        return Ok();
+        try
+        {
+            await FaceIDService.SetUserEnabledStatusAsync(userId, false);
+        
+            return Ok();
+            
+        }catch (UserNotFoundException e)
+        {
+            return NotFound(e.Message);
+        }
+        catch (Exception e)
+        {
+            Logger.Error(e, "Error enabling user faceid status");
+            return StatusCode(500, "Internal server error");
+        }
 
     }