+++ src/DAL/Entities/FaceIDUser.cs
@@ -14,7 +14,7 @@ public partial class FaceIDUser
     public int LastUpdateUserId { get; set; }
     
     public virtual User? LastUpdateUser { get; set; } = null!;
-    public virtual User User { get; set; }
-    
+    public virtual User User { get; set; } = null!;
+
     public virtual ICollection<BiometricTransaction> BiometricTransactions { get; set; } = new List<BiometricTransaction>();
 }