+++ src/Model/Plugins/PluginDll.cs
@@ -2,7 +2,7 @@ namespace Model.Plugins;
 
 public class PluginDll
 {
-    public string Name { get; set; }
-    public string Path { get; set; }
-    public string Type { get; set; }
+    public string Name { get; set; } = string.Empty;
+    public string Path { get; set; } = string.Empty;
+    public string Type { get; set; } = string.Empty;
 }